D4
Today we were playing games. The object was to finish the given features of product. Every feature has its own value points but also level of difficulty. Measuring current situation with divided teams and searching for best solution. It was fun and also very outstanding in order to understand processes happening between sprints and teams.