D6
Simon Sinek at Ted Talks. At this point of searching and analysing values and methods of Agile, we came to his video explaining the question "why?". I don't want to spoiler it, I can't but recommend to all of you. It all starts with "why?". If your why is not enough, then you will start to search for "because".