D20
Me and my friends have decided to sign for "Game Jam". It's contest for game developers straight for 2 days. Goal of this contest is to popularize and encourage game development in our country. As I'm on the way of learning Agile methods, i think i can integrate some methods and see it through practice.